+++ server/src/main/java/com/krillsson/sysapi/logaccess/windowseventlog/WindowsServiceCommandResult.kt
@@ -3,5 +3,6 @@ package com.krillsson.sysapi.logaccess.windowseventlog
 sealed class WindowsServiceCommandResult {
     object Success : WindowsServiceCommandResult()
     object Unavailable : WindowsServiceCommandResult()
+    object Disabled : WindowsServiceCommandResult()
     data class Failed(val error: Throwable) : WindowsServiceCommandResult()
 }